authentification_app/requirements.txt
asgiref==3.7.2
crispy-tailwind==0.5.0
Django==4.2.6
django-crispy-forms==2.1
django-webpack-loader==2.0.1
nodeenv==1.8.0
Pillow==10.1.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
